more test update, removed child class
