+++ tests/soil_crop_tests/soil_tests/phosphorus_cycling_tests/test_manure.py
@@ -12,128 +12,8 @@ from RUFAS.routines.field.soil.soil_data import SoilData
 from RUFAS.routines.field.soil.phosphorus_cycling.manure import Manure
 
 
-
-
-
-@pytest.mark.parametrize(
-    "phosphorus,rate,temp_factor,moisture_factor",
-    [
-        (25, 0.1, 0.33, 0.55),
-        (33, 0.01, 0.65, 0.78),
-        (21, 0.0025, 0.3423, 0.7768),
-        (0, 0.01, 0.012, 0.23),
-        (23, 0.0025, -0.13, 0.332),
-        (41, 0.1, -0.19, 0.0),
-    ],
-)
-def test_determine_mineralized_surface_phosphorus(
-    phosphorus: float, rate: float, temp_factor: float, moisture_factor: float
-) -> None:
-    """Tests that the correct amount of mineralized phosphorus is calculated."""
-    observed = Manure.determine_mineralized_surface_phosphorus(phosphorus, rate, temp_factor, moisture_factor)
-    expected = min(phosphorus, max(0.0, phosphorus * rate * min(temp_factor, moisture_factor)))
-    assert observed == expected
-
-
-@pytest.mark.parametrize(
-    "ratio,phosphorus",
-    [
-        (0.88, 26),
-        (0.212, 12.13),
-        (0.0, 30.21),
-        (0.441, 0.0),
-        (0.0, 0.0),
-    ],
-)
-def test_determine_assimilated_phosphorus_amount(ratio: float, phosphorus: float) -> None:
-    """Tests that the correct amount of phosphorus assimilated into the soil is calculated."""
-    observed = Manure.determine_assimilated_phosphorus_amount(ratio, phosphorus)
-    expected = max(0.0, ratio * phosphorus)
-    expected = min(phosphorus, expected)
-    assert observed == expected
-
-
 # --- Helper method tests ---
-@pytest.mark.parametrize(
-    "amount_phosphorus,field_size",
-    [
-        (100, 3.1),
-        (25.6, 2),
-        (66.23, 1.88),
-    ],
-)
-def test_add_infiltrated_phosphorus_to_soil(amount_phosphorus: float, field_size: float) -> None:
-    """Test that methods are called correctly on correct layers of soil profile."""
-    data = SoilData(field_size=field_size)
-    incorp = Manure(data)
-    with patch(
-        "RUFAS.routines.field.soil.layer_data.LayerData.add_to_labile_phosphorus",
-        new_callable=PropertyMock,
-    ) as mocked_add_to_labile_phosphorus:
-        incorp._add_infiltrated_phosphorus_to_soil(amount_phosphorus, field_size)
-        assert mocked_add_to_labile_phosphorus.call_count == 2
 
-
-@pytest.mark.parametrize(
-    "rain,runoff,area",
-    [
-        (13, 4, 1.8),
-        (12, 1.8, 2.1),
-        (14, 12.2, 3.4),
-        (4.2, 0, 2.4),
-    ],
-)
-def test_leach_and_update_phosphorus_pools(rain: float, runoff: float, area: float) -> None:
-    """Tests that the update subroutine for phosphorus pools in Manure correctly calls methods and sets attributes."""
-    data = SoilData(
-        machine_manure=ManurePool(
-            manure_dry_mass=1000,
-            manure_field_coverage=0.86,
-            water_extractable_inorganic_phosphorus=200,
-            water_extractable_organic_phosphorus=90,
-        ),
-        grazing_manure=ManurePool(
-            manure_dry_mass=800,
-            manure_field_coverage=0.78,
-            water_extractable_inorganic_phosphorus=125,
-            water_extractable_organic_phosphorus=70,
-        ),
-        field_size=area,
-    )
-    incorp = Manure(data)
-
-    incorp._determine_phosphorus_leached_from_surface = MagicMock(
-        return_value={
-            "new_phosphorus_pool_amount": 30,
-            "infiltrated_phosphorus": 25,
-            "runoff_phosphorus": 20,
-        }
-    )
-    incorp._add_infiltrated_phosphorus_to_soil = MagicMock()
-
-    incorp._leach_and_update_phosphorus_pools(rain, runoff, area)
-
-    leached_calls = [
-        call(rain, runoff, area, 1000, 0.86, 90, True),
-        call(rain, runoff, area, 1000, 0.86, 200, False),
-        call(rain, runoff, area, 800, 0.78, 70, True),
-        call(rain, runoff, area, 800, 0.78, 125, False),
-    ]
-    incorp._determine_phosphorus_leached_from_surface.assert_has_calls(leached_calls)
-    infiltrated_calls = [call(25, area), call(25, area), call(25, area), call(25, area)]
-    incorp._add_infiltrated_phosphorus_to_soil.assert_has_calls(infiltrated_calls)
-    assert incorp.data.machine_manure.water_extractable_organic_phosphorus == 30
-    assert incorp.data.machine_manure.water_extractable_inorganic_phosphorus == 30
-    assert incorp.data.machine_manure.organic_phosphorus_runoff == 20
-    assert incorp.data.machine_manure.inorganic_phosphorus_runoff == 20
-    assert incorp.data.machine_manure.annual_runoff_manure_organic_phosphorus == 20
-    assert incorp.data.machine_manure.annual_runoff_manure_inorganic_phosphorus == 20
-    assert incorp.data.grazing_manure.water_extractable_organic_phosphorus == 30
-    assert incorp.data.grazing_manure.water_extractable_inorganic_phosphorus == 30
-    assert incorp.data.grazing_manure.organic_phosphorus_runoff == 20
-    assert incorp.data.grazing_manure.inorganic_phosphorus_runoff == 20
-    assert incorp.data.grazing_manure.annual_runoff_manure_organic_phosphorus == 20
-    assert incorp.data.grazing_manure.annual_runoff_manure_inorganic_phosphorus == 20
 
 
 @pytest.mark.parametrize(
